feat: move domain logic to pkg/ folder
