+++ pkg/note/note.go
